Created IPEDS dictionary sort
Created a Python file that sorts the IPEDS data dictionaries into their
corresponding year in the codebooks folder. Slightly altered the Python
file that deletes duplicates, SPSS, and SAS files to allow for
duplicates greater than 9, and unzipped SPSS/SAS files. Added a
verification to the IPEDS year function that makes sure two years are
consecutive if the four digits are not a possible four digit year.
Added a list of files that must be manually sorted to their years to
the IPEDS year function. Added a check on the file that sorts data to
years that the file is not a dictionary.