Edited IPEDS create dta
Edited the Python files that create dta files from the csv files
downloaded from the IPEDS website. Added a shebang and header. Changed
the location of the IPEDS data directory. Rerouted the log files to the
log files folder. Added an argument to the create dta function that
points to the Stata executable path.