Edited IPEDS insheet error files
Edited the Python files that make a list of all the do files that
encountered an error while creating a dta file from the csv file
downloaded from the IPEDS website. Added a shebang and header. Changed
the location of the IPEDS directory. Renamed the log contains error
function to log error function.