Added Windows support to create dta
Added Windows support to the create dta file by looking in the standard
Stata install location for a copy of Stata. Changed the linux Stata
directory to search for a stata folder in the standard Stata
installation location. Edited the Mac support to remove any explicit
file separators. Made the stata_executable_path variable optional in
the create dta function.