Accounted for new file structure
Added an if condition that makes sure the folder is named a year so
that only files downloaded from IPEDS are edited and deleted. This is
to account for the fact that there are manually created do files in the
do files folder that would have been deleted without the if condition.