Edited do files and year function
Changed the year function to account for the 1999-2000 two digit year.
Changed the do file insheet data path to the raw data folder so that
the directory can be changed to the clean data folder so that the file
is saved with the clean data. Keeping the codebooks unzipped and
deleting the codebooks zips.