Edited which files are deleted at end
Changed the method in which files are deleted after the creation of dta
datasets so that only the folders created during the process are
checked. Keeping the unzipped codebooks files instead of the zipped
codebooks files because they will be referenced often. Added a line of
code to the create dta function that creates the clean data folder if
it doesn’t already exist.